+++ BE/src/main/java/com/codesquad/issue/domain/Issue/Issue.java
@@ -1,7 +1,8 @@
-package com.codesquad.issue.domain;
+package com.codesquad.issue.domain.Issue;
 
 import static java.time.LocalDateTime.now;
 
+import com.codesquad.issue.domain.Account;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
