+++ BE/src/main/java/com/codesquad/issue/domain/Issue/IssueRepository.java
@@ -1,6 +1,6 @@
-package com.codesquad.issue;
+package com.codesquad.issue.domain.Issue;
 
-import com.codesquad.issue.domain.Issue;
+import com.codesquad.issue.domain.Issue.Issue;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
